requirements.txt
Django==2.0.7
django-memoize==2.1.1
pytz==2018.5
